# app/api/customers/[id]/blacklist/route.ts
import { NextRequest, NextResponse } from "next/server";

export async function POST(request: NextRequest, { params }: { params: { id: string } }) {
    
    const customerId = Number(params.id);
    if(isNaN(customerId) || customerId <= 0){
         return NextResponse.json(
            { error: 'Invalid customer id'},
            { status: 400 }
        );
    }
    
    try{
        const body = await request.json();
        const response = await fetch(`http://139.59.140.232/api/customers/${customerId}/blacklist`, {
        method: 'POST',
        headers: {
            'Content-Type': 'application/json',
            'Authorization': request.headers.get('Authorization') || '',
        },
        body: JSON.stringify(body)
        });

        if(!response.ok){
            const errorData = await response.json().catch(() => null);
            return NextResponse.json({ error: errorData?.error || 'Failed to add customer to blacklist' }, { status: response.status });
        }

        const data = await response.json();
        return NextResponse.json(data);

    } catch(error){
        console.error('Error adding customer to blacklist:', error);
         return NextResponse.json(
      { error: 'Internal server error', details: error instanceof Error ? error.message : 'Unknown error' },
      { status: 500 }
    );
    }
}

export async function DELETE(request: NextRequest, { params }: { params: { id: string } }) {
    
    const customerId = Number(params.id);
    if(isNaN(customerId) || customerId <= 0){
         return NextResponse.json(
            { error: 'Invalid customer id'},
            { status: 400 }
        );
    }
    
    try{
        const response = await fetch(`http://139.59.140.232/api/customers/${customerId}/blacklist`, {
        method: 'DELETE',
        headers: {
            'Authorization': request.headers.get('Authorization') || '',
        },
        });

        if(!response.ok){
            const errorData = await response.json().catch(() => null);
            return NextResponse.json({ error: errorData?.error || 'Failed to remove customer from blacklist' }, { status: response.status });
        }

        const data = await response.json();
        return NextResponse.json(data);

    } catch(error){
        console.error('Error removing customer from blacklist:', error);
         return NextResponse.json(
      { error: 'Internal server error', details: error instanceof Error ? error.message : 'Unknown error' },
      { status: 500 }
    );
    }
}